pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arxiv-rag-v1"
version = "0.1.0"
description = "arXiv LLM 논문 RAG 시스템 - 포트폴리오 프로젝트"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "ajh428" }
]
keywords = ["rag", "llm", "arxiv", "embedding", "search"]

dependencies = [
    # Core
    "arxiv>=2.1.0",
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # HTTP & Async
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "tenacity>=8.0.0",

    # Parsing
    "marker-pdf>=1.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "pylatexenc>=2.10",

    # Embedding
    "FlagEmbedding>=1.2.0",
    "openai>=1.0.0",
    "tiktoken>=0.7.0",

    # API & UI
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "streamlit>=1.32.0",

    # AI
    "google-generativeai>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
